+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-03-10
+### Added
+- unsupervised learning model with Amazon SageMaker Random Cut Forest
+- data upsampling techniques like SMOTE for addressing imbalanced data
+- More interpretation of results
+- supervised learning model with XGBoost
+- Amazon API Gateway entrypoint
+
+### Removed
+- Supervised learning model with SageMaker linear learner
+
+
 ## [1.0.0] - 2019-05-16
 ### Added
 - initial checkin